Move playbook tasks to role
In preparation for other refactors.
This will allow for setting some default values for vars, so that the existing playbooks still run as they did previously without need to update the scripts running them.
having the role will also allow to use the file/ and templates/ sub-directories, and add some additional configurability to the role.